+++ Source/IGListStackedSectionController.m
@@ -16,6 +16,32 @@
 
 #import "IGListSectionControllerInternal.h"
 
+@interface UICollectionViewCell (IGListStackedSectionController)
+@end
+@implementation UICollectionViewCell (IGListStackedSectionController)
+
+static void * kStackedSectionControllerKey = &kStackedSectionControllerKey;
+
+- (void)ig_setStackedSectionController:(id)stackedSectionController {
+    objc_setAssociatedObject(self, kStackedSectionControllerKey, stackedSectionController, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
+
+- (id)ig_stackedSectionController {
+    return objc_getAssociatedObject(self, kStackedSectionControllerKey);
+}
+
+static void * kStackedSectionControllerIndexKey = &kStackedSectionControllerIndexKey;
+
+- (void)ig_setStackedSectionControllerIndex:(NSInteger)stackedSectionControllerIndex {
+    objc_setAssociatedObject(self, kStackedSectionControllerIndexKey, @(stackedSectionControllerIndex), OBJC_ASSOCIATION_ASSIGN);
+}
+
+- (NSInteger)ig_stackedSectionControllerIndex {
+    return [objc_getAssociatedObject(self, kStackedSectionControllerIndexKey) integerValue];
+}
+
+@end
+
 @implementation IGListStackedSectionController
 
 - (instancetype)initWithSectionControllers:(NSArray <IGListSectionController<IGListSectionType> *> *)sectionControllers {
@@ -30,6 +56,7 @@
 
         self.displayDelegate = self;
         self.scrollDelegate = self;
+        self.workingRangeDelegate = self;
 
         [self reloadData];
     }
@@ -284,6 +311,10 @@
     IGListSectionController<IGListSectionType> *childSectionController = [self sectionControllerForObjectIndex:index];
     const NSUInteger localIndex = [self localIndexForSectionController:childSectionController index:index];
 
+    // update the assoc objects for use in didEndDisplay
+    [cell ig_setStackedSectionController:childSectionController];
+    [cell ig_setStackedSectionControllerIndex:localIndex];
+
     NSCountedSet *visibleSectionControllers = self.visibleSectionControllers;
     id<IGListDisplayDelegate> displayDelegate = [childSectionController displayDelegate];
 
@@ -296,8 +327,9 @@
 }
 
 - (void)listAdapter:(IGListAdapter *)listAdapter didEndDisplayingSectionController:(IGListSectionController<IGListSectionType> *)sectionController cell:(UICollectionViewCell *)cell atIndex:(NSInteger)index {
-    IGListSectionController<IGListSectionType> *childSectionController = [self sectionControllerForObjectIndex:index];
-    const NSUInteger localIndex = [self localIndexForSectionController:childSectionController index:index];
+    const NSUInteger localIndex = [cell ig_stackedSectionControllerIndex];
+    IGListSectionController<IGListSectionType> *childSectionController = [cell ig_stackedSectionController];
+
     NSCountedSet *visibleSectionControllers = self.visibleSectionControllers;
     id<IGListDisplayDelegate> displayDelegate = [childSectionController displayDelegate];
 
@@ -329,6 +361,20 @@
 - (void)listAdapter:(IGListAdapter *)listAdapter didEndDraggingSectionController:(IGListSectionController<IGListSectionType> *)sectionController willDecelerate:(BOOL)decelerate {
     for (IGListSectionController<IGListSectionType> *childSectionController in self.sectionControllers) {
         [[childSectionController scrollDelegate] listAdapter:listAdapter didEndDraggingSectionController:childSectionController willDecelerate:decelerate];
+    }
+}
+
+#pragma mark - IGListWorkingRangeDelegate
+
+- (void)listAdapter:(IGListAdapter *)listAdapter sectionControllerWillEnterWorkingRange:(IGListSectionController<IGListSectionType> *)sectionController {
+    for (IGListSectionController<IGListSectionType> *childSectionController in self.sectionControllers) {
+        [[childSectionController workingRangeDelegate] listAdapter:listAdapter sectionControllerWillEnterWorkingRange:childSectionController];
+    }
+}
+
+- (void)listAdapter:(IGListAdapter *)listAdapter sectionControllerDidExitWorkingRange:(IGListSectionController<IGListSectionType> *)sectionController {
+    for (IGListSectionController<IGListSectionType> *childSectionController in self.sectionControllers) {
+        [[childSectionController workingRangeDelegate] listAdapter:listAdapter sectionControllerDidExitWorkingRange:childSectionController];
     }
 }